+++ .gitignore
@@ -164,5 +164,5 @@ tags
 
 # DS_Store (MacOS)
 .DS_Store
+# RL pipelines may produce mp4 outputs
 *.mp4
-hub/*
