+++ scripts/convert_models_diffuser_to_diffusers.py
